requirements.txt
pyramid
requests
Pillow
numpy